added generic configuration checking for inputs, instead of doing it manually in every subclass
    remove checkConfiguration(Configuration) method and replaced it with setConfiguration(config) + checkConfiguration() to maintain sanity

+++ dataset/GitHub_Java/Graylog2.graylog2-server/graylog2-plugin-interfaces/src/main/java/org/graylog2/plugin/configuration/Configuration.java
@@ -28,7 +28,11 @@ import org.slf4j.LoggerFactory;
 
 import java.util.Map;
 
+/**
+ * @author Lennart Koopmann <lennart@torch.sh>
+ */
 public class Configuration {
+
     private static final Logger LOG = LoggerFactory.getLogger(Configuration.class);
 
     private final Map<String, Object> source;
@@ -75,7 +79,7 @@ public class Configuration {
         strings.put(key, value);
     }
 
-    public int getInt(String key) {
+    public long getInt(String key) {
         return ints.get(key);
     }
